app/requirements.txt
Flask
Flask-Bootstrap
flask-mongoengine 
flask-nav
Flask-WTF
python-docx